Rename some crash related methods and improve debugging.
Using a debugger with the thread watcher active is kind of a pain because it
generates crashes that cannot be easily ignored after the fault. So generate
a debug break instead of an access violation.

Also, it is misleading to inspect crash dumps that have NullPointerCrash() on
the stack when the code doesn't actually crash.

BUG=

Review URL: https://codereview.chromium.org/477023002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@289866 4ff67af0-8c30-449e-8e8b-ad334ec8d88c