Log NTP loading time separately for local, Google and other NTPs.
We currently use the same UMA metric to log the loading time of different kinds of NTP:
- The local NTP
- The Google app-cached NTP
- NTP from other providers

This CL separates these in three UMA metric using histogram suffixes.

In addition, this CL brings back the Tab.NewTab* UMA metrics to the public histograms.xml in accordance with the practice that metrics in the public chromium repository should be in the public histograms.xml.

Review URL: https://codereview.chromium.org/474053002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@289946 4ff67af0-8c30-449e-8e8b-ad334ec8d88c