Fix webui cert viewer showing wrong cert chain on NSS and no chain on OpenSSL.
The webui cert viewer was still using x509_certificate_model::GetCertChainFromCert instead of X509Certificate::GetIntermediateCertificates. This was fixed in the gtk cert viewer in r135231, but the webui cert viewer was missed. (x509_certificate_model::GetCertChainFromCert was not implemented on OpenSSL, and on NSS the chain it returned may differ from the the chain that was actually used.)

BUG=77757,338887

Review URL: https://codereview.chromium.org/376753002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@289994 4ff67af0-8c30-449e-8e8b-ad334ec8d88c