Speculative fix for BrowserOptionsOverlayWebUITest.testNavigationInBackground
BrowserOptionsOverlayWebUITest.testNavigationInBackground fails on DrMemory
bots:
http://build.chromium.org/p/chromium.memory.fyi/builders/Windows%20Browser%20%28DrMemory%20full%29%20%2811%29/builds/476

This changes the test to wait if it detects that the appropriate message has not yet been
sent to the uber page. This ensures that we are only testing the finalized state of the
page.  If the test times out, that would indicate that the problem lies elsewhere.

BUG=403627
R=estade@chromium.org

Review URL: https://codereview.chromium.org/476573002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@290049 4ff67af0-8c30-449e-8e8b-ad334ec8d88c