Track Windows registry-sideloaded extensions.
This adds a creation flag for extensions which tracks if they
were installed from a potentially untrusted source.

The flag is currently set for extensions which are installed
via the Windows registry from a CRX on the local filesystem.

No enforcement action is being taken yet and no additional metrics are being collected.

BUG=325469

Review URL: https://codereview.chromium.org/479513004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@290122 4ff67af0-8c30-449e-8e8b-ad334ec8d88c