Add command-line flag for tests to bypass the screenshare window picker.
This flag will make it possible for external tests to bypass the dialog
where you select what to screen share. When this flag is passed on
startup and an extension asks to use the desktop capture API, it will
auto-bypass the dialogue if the target shows up in window picker.

This is assumed to be safe because the flag would have to be passed
each time Chrome is started up, rather than being an about:flags flag.

BUG=392425

Review URL: https://codereview.chromium.org/378273003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@290249 4ff67af0-8c30-449e-8e8b-ad334ec8d88c