Add wifi AP info to system profile metrics
Report wifi AP info to UMA system profile information stats for connected APs. The data will not be collected for the APs that contain "_nomap" in their SSID. The stats will be display/analyze in the dashboard that will be created with future CLs.

BUG=chromium:378892
TEST=Use NET_LOG messages to verify the AP vendor information. Will
also be verified in the future CLs.

Review URL: https://codereview.chromium.org/328793002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@290350 4ff67af0-8c30-449e-8e8b-ad334ec8d88c