Postpone keep alive reset for the app list.
This CL fixes an issue where dismissing the app list with no other browsers
would cause a CHECK fail in the event processing system. This was happening
because the closing of the app list would cause the event dispatcher to get
destroyed. This has been fixed by posting a task to reset the keep alive
which gives the event an opportunity to clean up.

BUG=395937

Review URL: https://codereview.chromium.org/484593002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@290488 4ff67af0-8c30-449e-8e8b-ad334ec8d88c