MacViews: Lint types in AppListControllerAppInfoDialogBrowserTest
This is part of getting browser_tests compiling on MacViews.

GetAllOwnedWidgets takes a NativeView, so a different argument is needed
on Mac. Also changes AppListControllerAppInfoDialogBrowserTest to be
just a typedef.

The test is still disabled. This CL doesn't change that, but the root
cause is now known (crrev/477923005 has this test, re-enabled and
passing passing on the bots).

BUG=399191

Review URL: https://codereview.chromium.org/483673002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@290489 4ff67af0-8c30-449e-8e8b-ad334ec8d88c