Fix network configurations that use the system CA certs.
Before, for EAP, the network configuration UI sent a list with an empty string to Shill instead of an empty list.
For VPNs, CACertPEM property was not sent at all and thus potentially not deleting a previously selected CA cert.

Now, in all cases, if no CA cert is selected, an empty list will be sent to Shill.

BUG=405067

Review URL: https://codereview.chromium.org/485083002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@290597 4ff67af0-8c30-449e-8e8b-ad334ec8d88c