Make DisplayInfoProvider an interface
Now is up to the component embedding the SystemInfo API to implement
the DisplayInfoProvider. The motivation for this change is we want to
move the SystemInfo API to extensions/ but DisplayInfoProvider depends
on ash/, which will link with ui/views/, and that is way too big.

The ultimate goal is to have SystemInfo on the app_shell without increasing its
footprint considerably.

BUG=392842

Review URL: https://codereview.chromium.org/476103002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@291250 4ff67af0-8c30-449e-8e8b-ad334ec8d88c