android: Don't allow AutofillPopup on outside touch
By default, PopupWindows dismiss when the user touches outside the window. The only
way to avert this behavior when using a ListPopupWindow is to rely on a hidden API.

This manifested as the linked bug because each tap dismissed the popup and detached it from the anchor view. C++ didn't realize the popup was dismissed, and tried to update it, but there was no anchor view.

BUG=400601

Review URL: https://codereview.chromium.org/442933002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@291463 4ff67af0-8c30-449e-8e8b-ad334ec8d88c